Snapshot of develop 2018-05-13 (Internal repo commit 3f4d172)
* 2019-05-13 (3f4d172) Fix missing copyright (NoJenkins)
* 2019-05-13 (9d4d6b7) Merge pull request #109 in AV/opencpi from task--AV-4957-matchstiq-z1-getting-started-guide-gui-instructions-dont-work to v1.5.0
* 2019-05-09 (220b4e3) Merge pull request #108 in AV/opencpi from github-io-link-fix to v1.5.0
* 2019-05-09 (d94111e) Merge remote-tracking branch 'origin/documentation-v1.5.0' into v1.5.0
* 2019-05-09 (29f67e8) adding testing for ocpidev directory finding
* 2019-05-09 (d850917) initial remote container admin scripts
* 2019-05-08 (2ecb46b) Merge pull request #104 in AV/opencpi from hotfix-AV-5477-python3-executable to v1.5.0
* 2019-05-07 (55a8518) sync with 1.5, from commit 0da3f33583c, but only part
* 2019-05-07 (9e30aed) sync with 1.5., too complicated to cherry-pick
* 2019-05-07 (acdec88) cherry pick from 1.5:  b58863a0323
* 2019-05-07 (2622456) Merge pull request #106 in AV/opencpi from bugfix--hotfix--AV-5495-fix-edge_detector-use-of-others-keyword to v1.5.0
* 2019-05-07 (b2eddb0) PDF creation
* 2019-05-07 (5c02b6d) Misc post-1.5 cleanup items
* 2019-05-07 (41f1f66) AV-4974 AV-5254 Update fir_complex_sse_ts to best practices
* 2019-05-07 (d8208f7) Merge pull request #105 in AV/opencpi from hotfix--AV-5259-make-improvements-to-the-edgeDetector-primitive to v1.5.0
* 2019-05-06 (4950f24) AV-5490: Break out udev note into snippet
* 2019-05-06 (b5a918f) Squashed commit of the following (AV-5479):
* 2019-05-03 (892d5dd) Merge pull request #101 in AV/opencpi from hotfix--AV-5481-update-arg-to-view.py to v1.5.0
* 2019-05-03 (240c78f) Merge pull request #103 in AV/opencpi from hotfix--AV-5480-remove-unused-view-assignment-from-test-xml to v1.5.0
* 2019-05-03 (31f4fa8) Merge pull request #98 in AV/opencpi from hotfix--fix-export-paths-for-FSK-app-Makefile to v1.5.0
* 2019-05-03 (8cf1aeb) Merge pull request #99 in AV/opencpi from hotfix-python3-symlink-missing to v1.5.0
*   2019-05-02 (7c6547d) Merge pull request #36 in AV/opencpi from feature--AV-5050-convert-docker-MAIN-to-multistage to develop
|\
| *   2019-05-02 (ffd5a41) Merge remote-tracking branch 'origin/develop' into feature--AV-5050-convert-docker-MAIN-to-multistage
| |\
| |/
|/|
* | 2019-05-02 (7404700) Merge pull request #97 in AV/opencpi from Task--AV-5476-update-debugging-tools-guide-based-on-feedback-for-signaltap to v1.5.0
* | 2019-05-02 (9b8d6d4) Merge pull request #100 in AV/opencpi from hotfix-wrong-centos-minor to v1.5.0
* | 2019-05-02 (4371881) File moving in 1.5
| * 2019-05-02 (a22ce2d) AV-5050 Dockerfiles: More accurate comments, little fixes.
| *   2019-05-02 (1bd5120) Merge remote-tracking branch 'origin/develop' into feature--AV-5050-convert-docker-MAIN-to-multistage
| |\
| |/
|/|
* | 2019-05-01 (cbe9567) removing reference to resolved ticket
| * 2019-05-01 (9da73c7) AV-5050 Lint Dockerfiles, use sleeper instead of sleep in 1 place, add nice labels, re-delete 1 moot file.
| *   2019-05-01 (47d1364) Merge remote-tracking branch 'origin/develop' into feature--AV-5050-convert-docker-MAIN-to-multistage
| |\
| |/
|/|
* | 2019-05-01 (c3fd750) Fix some low-level copyright problems. NoJenkins
* | 2019-05-01 (17d7154) Merge pull request #96 in AV/opencpi from hotfix--AV-5475-Update-doc-version-in-data-sheets-for-1.5-release- to v1.5.0
* | 2019-05-01 (94b16e2) Merge pull request #95 in AV/opencpi from hotfix--remove-WARNING-snippets-from-assets_ts to v1.5.0
* | 2019-05-01 (0c24af6) Merge pull request #94 in AV/opencpi from hotfix--remove-warning-from-FSK_app-doc to v1.5.0
* | 2019-05-01 (0b8d5f5) Add note about -gh suffix to Dev Policy
* | 2019-05-01 (7f114d2) AV-5471 Update doc versions to 1.5
* | 2019-05-01 (17aeffe) Merge pull request #92 in AV/opencpi from hotfix--av-5466-bad-entity to v1.5.0
* | 2019-05-01 (ef32207) Updated documentation from 1.5 to develop
* | 2019-05-01 (2e6313c) Merge pull request #88 in AV/opencpi from documentation-v1.5.0-code-edits to v1.5.0
* | 2019-04-30 (0a9b10a) Merge pull request #91 in AV/opencpi from hotfix--AV-5467-remove-message-size-from-rx-app-oas to v1.5.0
* | 2019-04-30 (65c48e0) Merge pull request #90 in AV/opencpi from hotfix--AV-5411-Rx-app-16352-to-8192 to v1.5.0
* | 2019-04-30 (d454136) Merge pull request #89 in AV/opencpi from bugfix--AV-5461-CLONE-Remove-messagesize-assignment-fromFSK_dig_radio_ctrlr-OAS-XML to v1.5.0
* | 2019-04-30 (366f807) (no ticket) PDF builder: Add python3 and fix 'multilib' problem with installing libgcc
* |   2019-04-30 (fcbe9ba) Merge remote-tracking branch 'origin/feature--AV-4803-data-plane-deep-dive' into develop
|\ \
| * | 2019-04-02 (6fe3ee8) AV-5199 Incremental revision to SDP timing diagrams
| * | 2019-03-05 (f2211cf) AV-5199 Initial commit of draft SDP timing diagrams
|  /
* | 2019-04-30 (8c00e2a) AV-5448 - pulled doc generator from doc branch NoJenkins
* | 2019-04-29 (12256b3) removed erroneous comment
* |   2019-04-29 (8b7f660) Merge remote-tracking branch 'origin/bugfix--AV-4784-cast-to-int-for-numpy' into develop NoJenkins
|\ \
| * | 2019-03-05 (bcfb220) Script cleanups
| * | 2019-03-04 (bc5ec09) AV-4784 With numpy >= 1.12.0: Indexing with floats raises IndexError, e.g., a[0, 0.0].
|  /
* | 2019-04-26 (164da77) Add TODO NoJenkins
* | 2019-04-26 (dd696e3) Merge pull request #86 in AV/opencpi from hotfix--AV-5355-pass-RPM-RELEASE-param-to-IDE-builder to v1.5.0
* | 2019-04-26 (16c5716) Merge pull request #85 in AV/opencpi from hotfix-sim-failure-on-1.5 to v1.5.0
* | 2019-04-25 (fa99c3b) Merge pull request #84 in AV/opencpi from hotfix--AV-5444-ad9361_dac_test-failing-because-of-property-warning to v1.5.0
* | 2019-04-25 (8cf7bdb) Typo NoJenkins
* | 2019-04-25 (2518883) A few TODOs in docGen NoJenkins
* | 2019-04-25 (de7caa7) Merge pull request #82 in AV/opencpi from hotfix--AV-5440-ad9361_adc_test-failing-because-of-property-warning to v1.5.0
* | 2019-04-24 (38a987a) added comments
* | 2019-04-24 (b276530) bugfix for AV-5442
* | 2019-04-24 (36d9db2) Merge pull request #75 in AV/opencpi from hotfix--build-all-alst4-at-once to v1.5.0
* | 2019-04-24 (12f472b) Merge pull request #77 in AV/opencpi from hotfix--AV-5410-aci-related-issue-filename-property-too-long to v1.5.0
* | 2019-04-24 (a40e98e) Merge pull request #79 in AV/opencpi from hotfix--AV-5437-time-service-pcie-related-issue-with-ml605-and-capture_v2 to v1.5.0
* | 2019-04-24 (25f12ee) Merge pull request #81 in AV/opencpi from bugfix--AV-5422-ntpd-bad-address-when-trying-source-mysetup.sh to v1.5.0
* | 2019-04-23 (2fe5cf0) separating python code into reusable files, also added doctest
* | 2019-04-23 (65d5455) AV-5408-induced fixes - iq_imbalance_fixer.test/verify.py - removed float array indeces incompatible w/ numpy >= 1.12.0, made 'find nearest frequency in FFT result' calculations more accurate, removed redundant code, created new class useful for future calculations
* | 2019-04-23 (bdb7988) Merge pull request #78 in AV/opencpi from bugfix--AV-5438-update-alst4-time-service-clock-to-match-control-plane-clock to v1.5.0
* | 2019-04-23 (45e8a51) Merge pull request #68 in AV/opencpi from hotfix--v1.5.0-AV-5355-IDE-build-needs-an-env-var to v1.5.0
* | 2019-04-23 (7efc996) Merge pull request #76 in AV/opencpi from hotfix--AV-5424-remove-messagesize-assignment-from-fsk-oas-xml-for-fmcomms to v1.5.0
| * 2019-04-19 (c03f693) AV-5050 review feedback ; NoJenkins
| *   2019-04-19 (4c851a5) Merge branch 'feature--AV-5050-convert-docker-MAIN-to-multistage' of <internal server>/opencpi into feature--AV-5050-convert-docker-MAIN-to-multistage
| |\
|/ /
| * 2019-04-16 (f347f8d) AV-5050 Specify sub-version of CentOS, e.g CentOS6.10. Add missing includes. Review feedback. NoJenkins
| * 2019-04-03 (4d822bc) AV-5050 Convert MAIN-building dockerfile to 'multi-stage' ; NoJenkins
| * 2019-04-03 (6ef9a57) AV-5050 Convert MAIN-building dockerfile to 'multi-stage'
* 2019-04-18 (bd1f457) adding packages
* 2019-04-18 (7cb4176) cdg update - not final for 1.5
* 2019-04-18 (7ab28ef) docGen.py - significant changes

+++ projects/assets_ts/components/fir_complex_sse_ts.test/gen_lpf_taps.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python2
 # This file is protected by Copyright. Please refer to the COPYRIGHT file
 # distributed with this source distribution.
 #
@@ -38,6 +38,7 @@ def set_taps_lpf():
         print("Exit: cutoff frequency out of range (0,1)")
         sys.exit(1)
     max_tap = pow(2,int(os.environ.get("OCPI_TEST_COEFF_WIDTH_p"))-1)-1
+    #Num taps is different than the assets version, which requires a separate script
     length = int(os.environ.get("OCPI_TEST_NUM_TAPS_p"))
 
     result = np.array(np.zeros(length), dtype=float)
@@ -64,11 +65,11 @@ def set_taps_lpf():
     #set taps scale to integers
     taps = (np.int16)(result * max_tap / scale)
     #print (float)(sum(taps))/(float)(max_tap)
-    #print taps[0:np.ceil(length/2.0)]
+    #print taps[0:int(np.ceil(length/2.0))]
 
     fo = open(sys.argv[1], 'w')
     #print "\tName of the output file:", fo.name
-    for i in taps[0:np.ceil(length/2.0)]:
+    for i in taps[0:int(np.ceil(length/2.0))]:
         stringy = ''.join(str(i)+'\n')
         fo.write(stringy)
     fo.close()
